+++ src/Curiosity.DAL.EF/CuriosityDbTransaction.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore.Storage;
 
 namespace Curiosity.DAL.EF
@@ -7,6 +9,8 @@ namespace Curiosity.DAL.EF
     public class CuriosityDbTransaction : ICuriosityDataContextTransaction
     {
         private readonly IDbContextTransaction _efDbContextTransaction;
+
+        private bool _wasDisposed;
 
         public CuriosityDbTransaction(IDbContextTransaction efDbContextTransaction)
         {
@@ -24,17 +28,41 @@ namespace Curiosity.DAL.EF
         {
             _efDbContextTransaction.Commit();
         }
+        
+        /// <inheritdoc />
+        public Task CommitAsync(CancellationToken cancellationToken = default)
+        {
+            return _efDbContextTransaction.CommitAsync(cancellationToken);
+        }
 
         /// <inheritdoc />
         public void Rollback()
         {
             _efDbContextTransaction.Rollback();
         }
+        
+        /// <inheritdoc />
+        public Task RollbackAsync(CancellationToken cancellationToken = default)
+        {
+            return _efDbContextTransaction.RollbackAsync(cancellationToken);
+        }
 
         /// <inheritdoc />
         public void Dispose()
         {
+            if (_wasDisposed) return;
+
+            _wasDisposed = true;
             _efDbContextTransaction.Dispose();
+        }
+
+        public ValueTask DisposeAsync()
+        {
+            if (_wasDisposed) return new ValueTask();
+
+            _wasDisposed = true;
+            
+            return _efDbContextTransaction.DisposeAsync();
         }
     }
 }