+++ src/Curiosity.DAL/ICuriosityDataContextFactory.cs
@@ -3,21 +3,27 @@ namespace Curiosity.DAL
     /// <summary>
     /// Data context factory (for <see cref="ICuriosityDataContext"/>) with typed returned value.
     /// </summary>
-    public interface ICuriosityDataContextFactory<out TReadWriteContext, out TReadOnlyContext> : ICuriosityDataContextFactory
-        where TReadWriteContext : ICuriosityDataContext
-        where TReadOnlyContext  : ICuriosityReadOnlyDataContext
+    public interface ICuriosityDataContextFactory<out TContext> : ICuriosityDataContextFactory
+        where TContext : ICuriosityDataContext
     {
         /// <summary>
-        /// Creates context <see cref="TReadWriteContext"/> for data access.
+        /// Creates context <see cref="TContext"/> for data access.
         /// </summary>
         /// <returns>New instance of data context</returns>
-        new TReadWriteContext CreateContext();
-        
+        new TContext CreateContext(bool isLoggingEnabled = false);
+    }
+    
+    /// <summary>
+    /// Data context factory (for <see cref="ICuriosityDataContext"/>) with typed returned value.
+    /// </summary>
+    public interface ICuriosityReadOnlyDataContextFactory<out TContext> : ICuriosityReadOnlyDataContextFactory
+        where TContext  : ICuriosityReadOnlyDataContext
+    {
         /// <summary>
-        /// Creates context <see cref="TReadOnlyContext"/> for data access  from replica (if no replica, main connection will be used).
+        /// Creates context <see cref="TContext"/> for read only data access (if no connection string found for read only access, main connection will be used).
         /// </summary>
         /// <returns>New instance of data context</returns>
-        new TReadOnlyContext CreateReplicaContext();
+        new TContext CreateReadOnlyContext(bool isLoggingEnabled = false);
     }
 
     /// <summary>
@@ -29,12 +35,18 @@ namespace Curiosity.DAL
         /// Creates context <see cref="ICuriosityDataContext"/> for data access.
         /// </summary>
         /// <returns>New instance of data context</returns>
-        ICuriosityDataContext CreateContext();
- 
+        ICuriosityDataContext CreateContext(bool isLoggingEnabled = false);
+    }
+    
+    /// <summary>
+    /// Data context factory (for <see cref="ICuriosityReadOnlyDataContext"/>).
+    /// </summary>
+    public interface ICuriosityReadOnlyDataContextFactory
+    {
         /// <summary>
-        /// Creates context <see cref="ICuriosityReadOnlyDataContext"/> for data access from replica (if no replica, main connection will be used).
+        /// Creates context <see cref="ICuriosityReadOnlyDataContext"/> for read only data access (if no connection string found for read only access, main connection will be used).
         /// </summary>
         /// <returns>New instance of data context</returns>
-        ICuriosityReadOnlyDataContext CreateReplicaContext();
+        ICuriosityReadOnlyDataContext CreateReadOnlyContext(bool isLoggingEnabled = false);
     }
 }