+++ src/Curiosity.DAL/ICuriosityDataContextTransaction.cs
@@ -1,11 +1,13 @@
 using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Curiosity.DAL
 {
     /// <summary>
     /// Transaction in work with <see cref="ICuriosityDataContext"/>.
     /// </summary>
-    public interface ICuriosityDataContextTransaction : IDisposable
+    public interface ICuriosityDataContextTransaction : IDisposable, IAsyncDisposable
     {
         /// <summary>
         /// Gets the transaction identifier.
@@ -21,10 +23,20 @@ namespace Curiosity.DAL
         /// Commits all changes made to the database in the current transaction.
         /// </summary>
         void Commit();
+        
+        /// <summary>
+        /// Commits all changes made to the database in the current transaction.
+        /// </summary>
+        Task CommitAsync(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Discards all changes made to the database in the current transaction.
         /// </summary>
         void Rollback();
+        
+        /// <summary>
+        /// Discards all changes made to the database in the current transaction.
+        /// </summary>
+        Task RollbackAsync(CancellationToken cancellationToken = default);
     }
 }