+++ src/Curiosity.DAL/ISqlExecutor.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Curiosity.DAL
@@ -15,11 +16,13 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         void Execute(
             ICuriosityDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = false);
+            bool ignoreNulls = false,
+            int? commandTimeoutSec = null);
 
         /// <summary>
         /// Executes SQl query.
@@ -28,11 +31,15 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task ExecuteAsync(
             ICuriosityDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = false);
+            bool ignoreNulls = false,
+            int? commandTimeoutSec = null,
+            CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Executes SQl query and returns value.
@@ -41,11 +48,13 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         T QuerySingleOrDefault<T>(
             ICuriosityReadOnlyDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = false);
+            bool ignoreNulls = false,
+            int? commandTimeoutSec = null);
 
         /// <summary>
         /// Executes SQl query and returns value.
@@ -54,11 +63,15 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task<T> QuerySingleOrDefaultAsync<T>(
             ICuriosityReadOnlyDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = false);
+            bool ignoreNulls = false,
+            int? commandTimeoutSec = null,
+            CancellationToken cancellationToken = default);
         
         /// <summary>
         /// Executes SQl query and returns collection of values.
@@ -67,11 +80,13 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         IEnumerable<T> QueryMany<T>(
             ICuriosityReadOnlyDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = false);
+            bool ignoreNulls = false,
+            int? commandTimeoutSec = null);
 
         /// <summary>
         /// Executes SQl query and returns collection of value.
@@ -80,11 +95,15 @@ namespace Curiosity.DAL
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task<IEnumerable<T>> QueryManyAsync<T>(
             ICuriosityReadOnlyDataContext context,
             string sqlTemplate, 
             IDictionary<string, object>? parameters = null,
-            bool ignoreNulls = true);
+            bool ignoreNulls = true,
+            int? commandTimeoutSec = null,
+            CancellationToken cancellationToken = default);
         
         /// <summary>
         /// Executes SQl query and returns collection of value.
@@ -92,10 +111,14 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="sqlTemplate">SQL query (can contains parameters placeholders)</param>
         /// <param name="parameters">Parameters for query</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task<IEnumerable<T>> QueryManyAsync<T>(
             ICuriosityReadOnlyDataContext context,
             string sqlTemplate, 
-            object? parameters = null);
+            object? parameters = null,
+            int? commandTimeoutSec = null,
+            CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Executes stored procedure and returns a result.
@@ -103,13 +126,13 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
         T ExecuteStoredProcedure<T>(
             ICuriosityDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
+            int? commandTimeoutSec = null,
             bool ignoreNulls = true);
 
         /// <summary>
@@ -118,14 +141,16 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task<T> ExecuteStoredProcedureAsync<T>(
             ICuriosityDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
-            bool ignoreNulls = true);
+            int? commandTimeoutSec = null,
+            bool ignoreNulls = true,
+            CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Executes stored procedure.
@@ -133,13 +158,13 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
         void ExecuteStoredProcedure(
             ICuriosityDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
+            int? commandTimeoutSec = null,
             bool ignoreNulls = true);
 
         /// <summary>
@@ -148,14 +173,16 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task ExecuteStoredProcedureAsync(
             ICuriosityDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
-            bool ignoreNulls = true);
+            int? commandTimeoutSec = null,
+            bool ignoreNulls = true,
+            CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Executes stored procedure and returns a result.
@@ -163,13 +190,13 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
         IEnumerable<T> ExecuteStoredProcedure<T>(
             ICuriosityReadOnlyDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
+            int? commandTimeoutSec = null,
             bool ignoreNulls = true);
 
         /// <summary>
@@ -178,13 +205,15 @@ namespace Curiosity.DAL
         /// <param name="context">Data context for query execution</param>
         /// <param name="procedureName">Stored procedure name</param>
         /// <param name="parameters">Parameters for stored procedure</param>
-        /// <param name="timeoutSec">Timeout for stored procedure</param>
+        /// <param name="commandTimeoutSec">Number of seconds before command execution timeout.</param>
         /// <param name="ignoreNulls">Null params will be ignored</param>
+        /// <param name="cancellationToken">Cancellation token.</param>
         Task<IEnumerable<T>> ExecuteStoredProcedureAsync<T>(
             ICuriosityReadOnlyDataContext context,
             string procedureName, 
             IDictionary<string, object>? parameters = null,
-            int? timeoutSec = null,
-            bool ignoreNulls = true);
+            int? commandTimeoutSec = null,
+            bool ignoreNulls = true,
+            CancellationToken cancellationToken = default);
     }
 }