+++ src/Curiosity.DAL.Dapper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2021-03-29
+
+### Added
+
+- Added timeout to all methods of `ISqlExecutor`.
+- Added cancellation token to all methods of `ISqlExecutor`.
+
 ## [1.0.1] - 2021-03-27
 
 ### Fixed