+++ src/Curiosity.DAL.EF/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.1.0] - 2021-03-29
+
+### Added
+
+- Added async methods for transactions.
+
+### Changed
+
+- Made `GetImmediateServerTimeUtcAsync` virtual to support new database types.
+
 ## [1.0.3] - 2021-03-27
 
 ### Fixed